Adjust CODEOWNERS to be sufficient to assign ownership
For repositories that are built in CD, we require that the ownership
of the repository must be a team, not just individual contributors, to
clearly signal ownership.

This adds the partner-enablement team, but keeps Murad and Jan Inge in
the list so they are also automatically explicitly assinged. Also,
Jan-Inge is not an official member of partner-enablement, so he needs
to be listed.